split tests into two packages
